[React] give player and host the safeMode Signed-off-by:Maxence Laurent <maxence.laurent@gmail.com>
